requirements.txt
PySide6>=6.6.0
boto3>=1.34.0
azure-identity>=1.15.0
azure-mgmt-storage>=21.1.0
azure-mgmt-resource>=23.0.0
google-cloud-storage>=2.14.0
google-auth>=2.27.0
reportlab>=4.0.0
